# src/lib/api/generated/schemas.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export const ItemSchema = {
    type: 'object',
    required: [
        'id',
        'name',
        'status',
        'category',
        'amount',
        'createdAt'
    ],
    properties: {
        id: {
            type: 'string'
        },
        name: {
            type: 'string'
        },
        description: {
            type: 'string'
        },
        status: {
            type: 'string',
            enum: [
                'active',
                'inactive',
                'archived'
            ]
        },
        category: {
            type: 'string'
        },
        amount: {
            type: 'number'
        },
        createdAt: {
            type: 'string',
            format: 'date-time'
        }
    }
} as const;

export const CreateItemRequestSchema = {
    type: 'object',
    required: [
        'name',
        'status',
        'category',
        'amount'
    ],
    properties: {
        name: {
            type: 'string'
        },
        description: {
            type: 'string'
        },
        status: {
            type: 'string',
            enum: [
                'active',
                'inactive',
                'archived'
            ]
        },
        category: {
            type: 'string'
        },
        amount: {
            type: 'number'
        }
    }
} as const;

export const UpdateItemRequestSchema = {
    type: 'object',
    properties: {
        name: {
            type: 'string'
        },
        description: {
            type: 'string'
        },
        status: {
            type: 'string',
            enum: [
                'active',
                'inactive',
                'archived'
            ]
        },
        category: {
            type: 'string'
        },
        amount: {
            type: 'number'
        }
    }
} as const;

export const ItemListResponseSchema = {
    type: 'object',
    required: [
        'items',
        'total',
        'page',
        'limit'
    ],
    properties: {
        items: {
            type: 'array',
            items: {
                $ref: '#/components/schemas/Item'
            }
        },
        total: {
            type: 'integer'
        },
        page: {
            type: 'integer'
        },
        limit: {
            type: 'integer'
        }
    }
} as const;