# src/lib/api/generated/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: 'http://localhost:3000/api' | (string & {});
};

export type Item = {
    id: string;
    name: string;
    description?: string;
    status: 'active' | 'inactive' | 'archived';
    category: string;
    amount: number;
    createdAt: string;
};

export type CreateItemRequest = {
    name: string;
    description?: string;
    status: 'active' | 'inactive' | 'archived';
    category: string;
    amount: number;
};

export type UpdateItemRequest = {
    name?: string;
    description?: string;
    status?: 'active' | 'inactive' | 'archived';
    category?: string;
    amount?: number;
};

export type ItemListResponse = {
    items: Array<Item>;
    total: number;
    page: number;
    limit: number;
};

export type GetItemsData = {
    body?: never;
    path?: never;
    query?: {
        page?: number;
        limit?: number;
        search?: string;
        status?: 'active' | 'inactive' | 'archived';
    };
    url: '/items';
};

export type GetItemsResponses = {
    /**
     * Successful list response
     */
    200: ItemListResponse;
};

export type GetItemsResponse = GetItemsResponses[keyof GetItemsResponses];

export type CreateItemData = {
    body: CreateItemRequest;
    path?: never;
    query?: never;
    url: '/items';
};

export type CreateItemResponses = {
    /**
     * Item created successfully
     */
    201: Item;
};

export type CreateItemResponse = CreateItemResponses[keyof CreateItemResponses];

export type DeleteItemData = {
    body?: never;
    path: {
        id: string;
    };
    query?: never;
    url: '/items/{id}';
};

export type DeleteItemResponses = {
    /**
     * Item deleted
     */
    204: void;
};

export type DeleteItemResponse = DeleteItemResponses[keyof DeleteItemResponses];

export type GetItemByIdData = {
    body?: never;
    path: {
        id: string;
    };
    query?: never;
    url: '/items/{id}';
};

export type GetItemByIdErrors = {
    /**
     * Item not found
     */
    404: unknown;
};

export type GetItemByIdResponses = {
    /**
     * Item found
     */
    200: Item;
};

export type GetItemByIdResponse = GetItemByIdResponses[keyof GetItemByIdResponses];

export type UpdateItemData = {
    body: UpdateItemRequest;
    path: {
        id: string;
    };
    query?: never;
    url: '/items/{id}';
};

export type UpdateItemResponses = {
    /**
     * Item updated
     */
    200: Item;
};

export type UpdateItemResponse = UpdateItemResponses[keyof UpdateItemResponses];